Fix recent_files config option automatically
The recent_files list has to be an python object serialization. This object
must be an array of strings. Python will save unicode strings and Python3
strings. Not following this rule make m64py crash either while loading the
recent_files information or when adding a new entry when a file is opened.