Change setup.py to Setuptools and refactor bits of code
Earlier today I got a warning from Pip saying that ability to uninstall
programs that were installed using the Distutils module is going to be
depricated. So I decided to update everything over to Setuptools, which
is the new Python standard. While doing so, I ran the script through
flake8 and pylint a bit and took care of warnings not dealing with line
length or the lack of docstrings. I also took a bit of creative license
when I felt comfortable doing so to make the script more readable.