Fix install error and convert README to reStructuredText
Early this week, GitHub user xX-Yoshi-Xx informed me that there was an
issue installing using my last commit. This was due to the script trying
to import a module from `src/m64py` before it was added to the path. So
I fixed that just by moving the import down a few lines.

I also rewrote the README in reStructuredText, since this is the
default Python markup format. I added some new content on using the
`requirements.txt`` file to install the Python dependencies as well as a
few other notes. I also did a few smaller things like making the
Changelog file in all caps like the other files.