Set correct resolution for high DPI displays (with vidext).
Resolution is still screwy in general for me (e.g. video is slightly too
high; window is resizable even though video plugin doesn't want it to
be...), but that happens regardless of whether I have high DPI enabled,
so it seems to be unrelated.

Not sure if this is the best way to do this.